# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DraftRiskItem": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "risk_type": {
            "title": "Risk Type",
            "type": "string"
          },
          "severity": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Severity",
            "type": "string"
          },
          "source_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Label"
          },
          "target_field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Field"
          }
        },
        "required": [
          "risk_type",
          "severity",
          "action"
        ],
        "title": "DraftRiskItem",
        "type": "object"
      },
      "DraftRiskReport": {
        "additionalProperties": false,
        "properties": {
          "badcase_violations": {
            "default": 0,
            "title": "Badcase Violations",
            "type": "integer"
          },
          "llm_auto_accepted_count": {
            "default": 0,
            "title": "Llm Auto Accepted Count",
            "type": "integer"
          },
          "must_not_auto_activate": {
            "default": true,
            "title": "Must Not Auto Activate",
            "type": "boolean"
          },
          "risk_count": {
            "title": "Risk Count",
            "type": "integer"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/DraftRiskItem"
            },
            "title": "Risks",
            "type": "array"
          }
        },
        "required": [
          "risk_count"
        ],
        "title": "DraftRiskReport",
        "type": "object"
      },
      "DraftSchema": {
        "additionalProperties": false,
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/DraftSchemaField"
            },
            "title": "Fields",
            "type": "array"
          },
          "must_not_auto_activate": {
            "default": true,
            "title": "Must Not Auto Activate",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sample_count": {
            "title": "Sample Count",
            "type": "integer"
          },
          "schema_id": {
            "title": "Schema Id",
            "type": "string"
          },
          "status": {
            "const": "draft",
            "default": "draft",
            "title": "Status",
            "type": "string"
          },
          "version": {
            "default": "0.1.0-draft",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "schema_id",
          "name",
          "sample_count",
          "fields"
        ],
        "title": "DraftSchema",
        "type": "object"
      },
      "DraftSchemaField": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "evidence_paths": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Paths",
            "type": "array"
          },
          "field_id": {
            "title": "Field Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "required_recommended": {
            "default": false,
            "title": "Required Recommended",
            "type": "boolean"
          },
          "review_required": {
            "default": false,
            "title": "Review Required",
            "type": "boolean"
          },
          "risk_flags": {
            "items": {
              "type": "string"
            },
            "title": "Risk Flags",
            "type": "array"
          },
          "source_evidence": {
            "items": {
              "type": "string"
            },
            "title": "Source Evidence",
            "type": "array"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "field_id",
          "name",
          "display_name",
          "type",
          "description",
          "source_evidence",
          "evidence_paths",
          "confidence"
        ],
        "title": "DraftSchemaField",
        "type": "object"
      },
      "DraftTemplate": {
        "additionalProperties": false,
        "properties": {
          "alias_rules": {
            "items": {
              "$ref": "#/components/schemas/TemplateDraftRule"
            },
            "title": "Alias Rules",
            "type": "array"
          },
          "must_not_auto_activate": {
            "default": true,
            "title": "Must Not Auto Activate",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "regex_suggestions": {
            "items": {
              "$ref": "#/components/schemas/RegexDraftSuggestion"
            },
            "title": "Regex Suggestions",
            "type": "array"
          },
          "schema_id": {
            "title": "Schema Id",
            "type": "string"
          },
          "status": {
            "const": "draft",
            "default": "draft",
            "title": "Status",
            "type": "string"
          },
          "template_id": {
            "title": "Template Id",
            "type": "string"
          },
          "version": {
            "default": "0.1.0-draft",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "template_id",
          "schema_id",
          "name"
        ],
        "title": "DraftTemplate",
        "type": "object"
      },
...
      "FieldCandidate": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "evidence_paths": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Paths",
            "type": "array"
          },
          "field_name": {
            "title": "Field Name",
            "type": "string"
          },
          "frequency": {
            "title": "Frequency",
            "type": "number"
          },
          "inferred_type": {
            "enum": [
              "string",
              "date",
              "amount",
              "identifier",
              "number",
              "boolean"
            ],
            "title": "Inferred Type",
            "type": "string"
          },
          "review_required": {
            "default": false,
            "title": "Review Required",
            "type": "boolean"
          },
          "risk_flags": {
            "items": {
              "type": "string"
            },
            "title": "Risk Flags",
            "type": "array"
          },
          "source_labels": {
            "items": {
              "type": "string"
            },
            "title": "Source Labels",
            "type": "array"
          },
          "value_examples": {
            "items": {
              "type": "string"
            },
            "title": "Value Examples",
            "type": "array"
          }
        },
        "required": [
          "field_name",
          "source_labels",
          "frequency",
          "inferred_type",
          "evidence_paths",
          "confidence"
        ],
        "title": "FieldCandidate",
        "type": "object"
      },
      "FieldDiscoveryResult": {
        "additionalProperties": false,
        "properties": {
          "field_candidates": {
            "items": {
              "$ref": "#/components/schemas/FieldCandidate"
            },
            "title": "Field Candidates",
            "type": "array"
          },
          "llm_auto_accepted_count": {
            "default": 0,
            "title": "Llm Auto Accepted Count",
            "type": "integer"
          },
          "sample_count": {
            "title": "Sample Count",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "sample_count",
          "field_candidates"
        ],
        "title": "FieldDiscoveryResult",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegexDraftSuggestion": {
        "additionalProperties": false,
        "properties": {
          "negative_examples": {
            "items": {
              "type": "string"
            },
            "title": "Negative Examples",
            "type": "array"
          },
          "pattern": {
            "title": "Pattern",
            "type": "string"
          },
          "positive_examples": {
            "items": {
              "type": "string"
            },
            "title": "Positive Examples",
            "type": "array"
          },
          "review_required": {
            "default": true,
            "title": "Review Required",
            "type": "boolean"
          },
          "target_field": {
            "title": "Target Field",
            "type": "string"
          }
        },
        "required": [
          "target_field",
          "pattern"
        ],
        "title": "RegexDraftSuggestion",
        "type": "object"
      },
...
      "SchemaDraftDiscoverRequest": {
        "additionalProperties": false,
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/UIRDocument-Input"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Documents",
            "type": "array"
          }
        },
        "required": [
          "documents"
        ],
        "title": "SchemaDraftDiscoverRequest",
        "type": "object"
      },
      "SchemaDraftExportResponse": {
        "additionalProperties": false,
        "properties": {
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "files": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Files",
            "type": "object"
          },
          "must_not_auto_activate": {
            "default": true,
            "title": "Must Not Auto Activate",
            "type": "boolean"
          },
          "sha256": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Sha256",
            "type": "object"
          }
        },
        "required": [
          "draft_id",
          "files",
          "sha256"
        ],
        "title": "SchemaDraftExportResponse",
        "type": "object"
      },
      "SchemaDraftGenerateRequest": {
        "additionalProperties": false,
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/UIRDocument-Input"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Documents",
            "type": "array"
          },
          "schema_id": {
            "title": "Schema Id",
            "type": "string"
          },
          "schema_name": {
            "title": "Schema Name",
            "type": "string"
          },
          "template_id": {
            "title": "Template Id",
            "type": "string"
          }
        },
        "required": [
          "documents",
          "schema_id",
          "schema_name",
          "template_id"
        ],
        "title": "SchemaDraftGenerateRequest",
        "type": "object"
      },
      "SchemaDraftPackage": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "discovery": {
            "$ref": "#/components/schemas/FieldDiscoveryResult"
          },
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "draft_report": {
            "$ref": "#/components/schemas/SchemaDraftReport"
          },
          "draft_schema": {
            "$ref": "#/components/schemas/DraftSchema"
          },
          "draft_template": {
            "$ref": "#/components/schemas/DraftTemplate"
          },
          "must_not_auto_activate": {
            "default": true,
            "title": "Must Not Auto Activate",
            "type": "boolean"
          },
          "risk_report": {
            "$ref": "#/components/schemas/DraftRiskReport"
          },
          "status": {
            "const": "draft",
            "default": "draft",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "draft_id",
          "created_at",
          "discovery",
          "draft_schema",
          "draft_template",
          "risk_report",
          "draft_report"
        ],
        "title": "SchemaDraftPackage",
        "type": "object"
      },
      "SchemaDraftReport": {
        "additionalProperties": false,
        "properties": {
          "badcase_violations": {
            "default": 0,
            "title": "Badcase Violations",
            "type": "integer"
          },
          "candidate_count": {
            "title": "Candidate Count",
            "type": "integer"
          },
          "deterministic": {
            "default": true,
            "title": "Deterministic",
            "type": "boolean"
          },
          "llm_auto_accepted_count": {
            "default": 0,
            "title": "Llm Auto Accepted Count",
            "type": "integer"
          },
          "must_not_auto_activate": {
            "default": true,
            "title": "Must Not Auto Activate",
            "type": "boolean"
          },
          "risk_count": {
            "title": "Risk Count",
            "type": "integer"
          },
          "sample_count": {
            "title": "Sample Count",
            "type": "integer"
          },
          "secret_leak_count": {
            "default": 0,
            "title": "Secret Leak Count",
            "type": "integer"
          },
          "source_doc_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Doc Ids",
            "type": "array"
          }
        },
        "required": [
          "sample_count",
          "candidate_count",
          "source_doc_ids",
          "risk_count"
        ],
        "title": "SchemaDraftReport",
        "type": "object"
      },
...
      "SourceAnchor": {
        "additionalProperties": false,
        "properties": {
          "bbox": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          }
        },
        "title": "SourceAnchor",
        "type": "object"
      },
      "TargetField": {
        "additionalProperties": false,
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "constraints": {
            "title": "Constraints",
            "type": "object"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "field_id": {
            "title": "Field Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "required": {
            "default": false,
            "title": "Required",
            "type": "boolean"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "field_id",
          "name",
          "display_name",
          "type"
        ],
        "title": "TargetField",
        "type": "object"
      },
      "TargetSchema": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/TargetField"
            },
            "title": "Fields",
            "type": "array"
          },
          "json_schema": {
            "title": "Json Schema",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "schema_id": {
            "title": "Schema Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "schema_id",
          "name",
          "version",
          "fields"
        ],
        "title": "TargetSchema",
        "type": "object"
      },
...
      "TemplateDraftRule": {
        "additionalProperties": false,
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "evidence_paths": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Paths",
            "type": "array"
          },
          "review_required": {
            "default": false,
            "title": "Review Required",
            "type": "boolean"
          },
          "risk_flags": {
            "items": {
              "type": "string"
            },
            "title": "Risk Flags",
            "type": "array"
          },
          "target_field": {
            "title": "Target Field",
            "type": "string"
          }
        },
        "required": [
          "target_field",
          "aliases",
          "confidence",
          "evidence_count",
          "evidence_paths"
        ],
        "title": "TemplateDraftRule",
        "type": "object"
      },
...
      "UIRAsset": {
        "additionalProperties": false,
        "properties": {
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "source_block_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Block Id"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "type",
          "path"
        ],
        "title": "UIRAsset",
        "type": "object"
      },
      "UIRBlock": {
        "additionalProperties": false,
        "properties": {
          "attributes": {
            "title": "Attributes",
            "type": "object"
          },
          "block_id": {
            "title": "Block Id",
            "type": "string"
          },
          "level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "source_anchor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceAnchor"
              },
              {
                "type": "null"
              }
            ]
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "block_id",
          "type"
        ],
        "title": "UIRBlock",
        "type": "object"
      },
      "UIRDocument-Input": {
        "additionalProperties": false,
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/UIRAsset"
            },
            "title": "Assets",
            "type": "array"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/UIRBlock"
            },
            "title": "Blocks",
            "type": "array"
          },
          "doc_id": {
            "title": "Doc Id",
            "type": "string"
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          },
          "normalization_records": {
            "items": {
              "type": "object"
            },
            "title": "Normalization Records",
            "type": "array"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UIRSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "uir_version": {
            "title": "Uir Version",
            "type": "string"
          }
        },
        "required": [
          "uir_version",
          "doc_id"
        ],
        "title": "UIRDocument",
        "type": "object"
      },
...
      "UIRSource": {
        "additionalProperties": false,
        "properties": {
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "upstream_agents": {
            "items": {
              "type": "string"
            },
            "title": "Upstream Agents",
            "type": "array"
          }
        },
        "required": [
          "source_type",
          "source_name"
        ],
        "title": "UIRSource",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SchemaPack Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/schema-drafts/discover": {
      "post": {
        "operationId": "discover_fields_api_v1_schema_drafts_discover_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchemaDraftDiscoverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldDiscoveryResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discover Fields",
        "tags": [
          "schema-drafts"
        ]
      }
    },
    "/api/v1/schema-drafts/generate": {
      "post": {
        "operationId": "generate_schema_draft_api_v1_schema_drafts_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchemaDraftGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDraftPackage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Schema Draft",
        "tags": [
          "schema-drafts"
        ]
      }
    },
    "/api/v1/schema-drafts/{draft_id}": {
      "get": {
        "operationId": "get_schema_draft_api_v1_schema_drafts__draft_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDraftPackage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Schema Draft",
        "tags": [
          "schema-drafts"
        ]
      }
    },
    "/api/v1/schema-drafts/{draft_id}/export": {
      "post": {
        "operationId": "export_schema_draft_api_v1_schema_drafts__draft_id__export_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDraftExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Schema Draft",
        "tags": [
          "schema-drafts"
        ]
      }
    },
    "/api/v1/schema-drafts/{draft_id}/validate": {
      "post": {
        "operationId": "validate_schema_draft_api_v1_schema_drafts__draft_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftRiskReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Schema Draft",
        "tags": [
          "schema-drafts"
        ]
      }
    },
...
    "/api/v1/schemas/{schema_id}": {
      "get": {
        "operationId": "get_schema_api_v1_schemas__schema_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "schema_id",
            "required": true,
            "schema": {
              "title": "Schema Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Schema",
        "tags": [
          "schemas"
        ]
      }
    },
...
  }
}